Adds option to dump LLVM IR
Introduces an optional `dump_ir` parameter to the `Expr` filter. When provided, the generated LLVM Intermediate Representation (IR) for the expression will be written to the specified file.

This feature aids in debugging, understanding, and optimizing the JIT-compiled expressions by allowing inspection of the generated code.

Also removes the Codacy configuration file.