Fix crash caused by unsequenced evaluation in createVideoFilter call
This commit fixes a critical crash that occurred when creating video
filters, manifesting as an invalid pointer (0x18) being passed as the
VSVideoInfo parameter to VSCore::isValidVideoFormat.

The crash was caused by unsequenced evaluation of function arguments
in the createVideoFilter call. The original code:

```cpp
    vsapi->createVideoFilter(out, "Expr", &d->vi, exprGetFrame,
                             exprFree, fmParallel, deps.data(),
                             deps.size(), d.release(), core);
```

contained both `&d->vi` and `d.release()` as arguments. In C++,
the evaluation order of function arguments is indeterminately
sequenced, meaning the compiler may evaluate them in any order.

If the compiler chose to evaluate `d.release()` before `&d->vi`:
1. d.release() transfers ownership and sets d's internal pointer to nullptr
2. &d->vi then attempts to access a member of the now-null unique_ptr
3. This results in &(nullptr->vi), which evaluates to the offset of
   the vi member within ExprData (0x18 bytes)

Fixes #5